1. Added jquery library file to local catalog 2. Fixed for RDB increment field

+++ src/main/java/com/iread/controller/IReadBookController.java
@@ -96,7 +96,7 @@ public class IReadBookController {
         String userName = principal.getName();
         User user = userService.getUserByName(userName);
 
-        ratingService.postBookRating(user, book, rate.intValue());
+        ratingService.postBookRating(user, book, rate);
         return HttpStatus.OK.getReasonPhrase();
     }
